Moving the validation class from Utility to the new Validation namescapes, adjusting docblocks and tests accordingly
