+++ Cake/Test/TestCase/Controller/ControllerTest.php
@@ -751,6 +751,7 @@ class ControllerTest extends TestCase {
  * @return void
  */
 	public function testValidateErrors() {
+		$this->markTestIncomplete('Need to revisit once models work again.');
 		ClassRegistry::flush();
 		$request = new Request('controller_posts/index');
 
@@ -776,6 +777,7 @@ class ControllerTest extends TestCase {
  * @return void
  */
 	public function testValidateErrorsOnArbitraryModels() {
+		$this->markTestIncomplete('Need to revisit once models work again.');
 		Configure::write('Config.language', 'eng');
 		$TestController = new TestController();
