+++ Cake/Validation/Validator.php
@@ -101,7 +101,7 @@ class Validator implements \ArrayAccess, \IteratorAggregate, \Countable {
  *
  * @param string $name [optional] The fieldname to fetch.
  * @param \Cake\Validation\ValidationSet $set The set of rules for field
- * @return Cake\Model\Validator\ValidationSet
+ * @return Cake\Validation\ValidationSet
  */
 	public function field($name, ValidationSet $set = null) {
 		if (empty($this->_fields[$name])) {
@@ -128,7 +128,7 @@ class Validator implements \ArrayAccess, \IteratorAggregate, \Countable {
 				return $this->_scopes[$name];
 			}
 			if ($name === 'default') {
-				return $this->_scopes[$name] = '\Cake\Utility\Validation';
+				return $this->_scopes[$name] = '\Cake\Validation\Validation';
 			}
 			return null;
 		}
@@ -140,7 +140,7 @@ class Validator implements \ArrayAccess, \IteratorAggregate, \Countable {
  * Sets the I18n domain for validation messages. This method is chainable.
  *
  * @param string $validationDomain The validation domain to be used.
- * @return Cake\Model\Validator
+ * @return Cake\Validation
  */
 	public function setValidationDomain($validationDomain) {
 		$this->_validationDomain = $validationDomain;
@@ -161,7 +161,7 @@ class Validator implements \ArrayAccess, \IteratorAggregate, \Countable {
  * Returns the rule set for a field
  *
  * @param string $field name of the field to check
- * @return Cake\Model\Validator\ValidationSet
+ * @return Cake\Validation\ValidationSet
  */
 	public function offsetGet($field) {
 		return $this->field($field);
@@ -171,7 +171,7 @@ class Validator implements \ArrayAccess, \IteratorAggregate, \Countable {
  * Sets the rule set for a field
  *
  * @param string $field name of the field to set
- * @param array|Cake\Model\Validator\ValidationSet $rules set of rules to apply to field
+ * @param array|Cake\Validation\ValidationSet $rules set of rules to apply to field
  * @return void
  */
 	public function offsetSet($field, $rules) {
